requirements.txt
Django==1.4.3
Fabric==1.4.3
South==0.7.6
Unipath==0.2.1
amqplib==1.0.2
anyjson==0.3.3
billiard==2.7.3.12
celery==3.0.6
distribute==0.6.27
dj-database-url==0.2.1
django-appconf==0.5
django-braces==0.2.1
-e git://github.com/3atmospheres/django-cache-panel.git@bbde3df47ea5fb5f58939aab810fbfdf557afd6c#egg=django_cache_panel-dev
django-celery==3.0.6
django-compressor==1.2
django-debug-toolbar==0.9.4
django-extensions==0.9
django-forms-bootstrap==2.0.3.post1
django-registration==0.8
django-social-auth==0.7.4
django-taggit==0.9.3
django-tastypie==0.9.11
gunicorn==0.14.6
httplib2==0.7.4
kombu==2.4.0
mimeparse==0.1.3
newrelic==1.10.0.28
oauth2==1.5.211
psycopg2==2.4.5
pycrypto==2.6
python-dateutil==1.5
python-openid==2.2.5
ssh==1.7.14
wsgiref==0.1.2
